feat(sindi): optimize SparseVector handling and improve sorting efficiency
1. Added mandatory SparseVector sorting by ID in ascending order within SINDI's Add and UpdateVector methods.

2. Removed redundant sorting logic in search_impl to reduce overhead.

3. Optimized is_subset_of_sparse_vector() by implementing a high-performance two-pointer algorithm.

4. Refactored sort_sparse_vector into a templated utility for compile-time ID/Value and Asc/Desc sorting.

5. Added SparseVector constructor for Vector<std::pair> with allocator support.

6. Fixed clang-tidy 'bugprone-integer-division' in test_index.cpp to prevent precision loss.

+++ tests/test_index.cpp
@@ -2144,7 +2144,7 @@ TestIndex::TestRecoverRemoveIndex(const IndexPtr& index,
                 }
             }
         }
-        float recall_removed = ((float)correct) / (base_num / 2);
+        float recall_removed = ((float)correct) / ((float)base_num / 2);
         REQUIRE(recall_removed >= 0.90);
 
         // [step 3] add <original data> into index again for recovery
